+++ src/request.ts
@@ -17,24 +17,32 @@ export interface Data {
 
 export interface PostRequestOptions extends RequestOptions {
     json?: Data;
+    appName?: string;
+    instanceId?: string;
 }
 
-export const post = (options: PostRequestOptions, cb) => {
-    if (!options.timeout) {
-        options.timeout = 10000;
-    }
-    return request.post(options, cb);
-};
-
-export const get = ({ url, etag, appName, instanceId }: GetRequestOptions, cb) => {
+export const post = ({ url, appName, timeout, instanceId }: PostRequestOptions, cb) => {
     const options = {
         url,
-        timeout: 10000,
+        timeout: timeout || 10000,
         headers: {
             'UNLEASH-APPNAME': appName,
             'UNLEASH-INSTANCEID': instanceId,
             'User-Agent': appName,
         },
+    };
+    return request.post(options, cb);
+};
+
+export const get = ({ url, etag, appName, timeout, instanceId }: GetRequestOptions, cb) => {
+    const options = {
+        url,
+        timeout: timeout || 10000,
+        headers: {
+            'UNLEASH-APPNAME': appName,
+            'UNLEASH-INSTANCEID': instanceId,
+            'User-Agent': appName
+        }
     };
     if (etag) {
         options.headers['If-None-Match'] = etag;